Rename fleet_orders.py to AIFleetOrder.py for review proposes.
